Separate utils into separate files
